pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "RiskSense-VLA"
version = "1.0.0"
description = "Hazard-Aware Predictive Vision-Language-Action system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "numpy>=1.26",
  "opencv-python>=4.9",
  "pillow>=10.0",
  "PyYAML>=6.0",
  "torch>=2.1",
  "torchvision>=0.16",
  "transformers>=4.48,<5",
  "peft>=0.10",
  "backoff>=2.2",
  "onnxruntime>=1.17",
  "pydantic>=2.6",
  "matplotlib>=3.8",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.5.0",
]
cuda = [
  "tensorrt>=10.0; platform_system == 'Linux'",
]
synthetic = [
  "diffusers>=0.29",
  "accelerate>=0.30",
]
open_vocab = [
  "ultralytics>=8.3",
  "timm>=1.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint.per-file-ignores]
"scripts/*.py" = ["E402"]
